+++ azuredevops/provider_test.go
@@ -19,7 +19,7 @@ func TestAzureDevOpsProvider_HasChildResources(t *testing.T) {
 		"azuredevops_build_definition",
 		"azuredevops_project",
 		"azuredevops_serviceendpoint_github",
-		"azuredevops_serviceendpoint_dockerhub",
+		"azuredevops_serviceendpoint_dockerregistry",
 		"azuredevops_serviceendpoint_azurerm",
 		"azuredevops_serviceendpoint_bitbucket",
 		"azuredevops_serviceendpoint_kubernetes",