+++ azuredevops/resource_serviceendpoint_dockerregistry_test.go
@@ -1,4 +1,4 @@
-// +build all resource_serviceendpoint_dockerhub
+// +build all resource_serviceendpoint_dockerregistry
 
 package azuredevops
 
@@ -22,11 +22,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-var dhTestServiceEndpointID = uuid.New()
-var dhRandomServiceEndpointProjectID = uuid.New().String()
-var dhTestServiceEndpointProjectID = &dhRandomServiceEndpointProjectID
+var dockerRegistryTestServiceEndpointID = uuid.New()
+var dockerRegistryRandomServiceEndpointProjectID = uuid.New().String()
+var dockerRegistryTestServiceEndpointProjectID = &dockerRegistryRandomServiceEndpointProjectID
 
-var dhTestServiceEndpoint = serviceendpoint.ServiceEndpoint{ //todo change
+var dockerRegistryTestServiceEndpoint = serviceendpoint.ServiceEndpoint{ //todo change
 	Authorization: &serviceendpoint.EndpointAuthorization{
 		Parameters: &map[string]string{
 			"username": "DH_TEST_username",
@@ -37,9 +37,9 @@ var dhTestServiceEndpoint = serviceendpoint.ServiceEndpoint{ //todo change
 		Scheme: converter.String("UsernamePassword"),
 	},
 	Data: &map[string]string{
-		"registrytype": "DockerHub",
+		"registrytype": "Others",
 	},
-	Id:          &dhTestServiceEndpointID,
+	Id:          &dockerRegistryTestServiceEndpointID,
 	Name:        converter.String("UNIT_TEST_CONN_NAME"),
 	Description: converter.String("UNIT_TEST_CONN_DESCRIPTION"),
 	Owner:       converter.String("library"), // Supported values are "library", "agentcloud"
@@ -52,30 +52,30 @@ var dhTestServiceEndpoint = serviceendpoint.ServiceEndpoint{ //todo change
  */
 
 // verifies that the flatten/expand round trip yields the same service endpoint
-func TestAzureDevOpsServiceEndpointDockerHub_ExpandFlatten_Roundtrip(t *testing.T) {
-	resourceData := schema.TestResourceDataRaw(t, resourceServiceEndpointDockerHub().Schema, nil)
-	flattenServiceEndpointDockerHub(resourceData, &dhTestServiceEndpoint, dhTestServiceEndpointProjectID)
+func TestAzureDevOpsServiceEndpointDockerRegistry_ExpandFlatten_Roundtrip(t *testing.T) {
+	resourceData := schema.TestResourceDataRaw(t, resourceServiceEndpointDockerRegistry().Schema, nil)
+	flattenServiceEndpointDockerRegistry(resourceData, &dockerRegistryTestServiceEndpoint, dockerRegistryTestServiceEndpointProjectID)
 
-	serviceEndpointAfterRoundTrip, projectID, err := expandServiceEndpointDockerHub(resourceData)
+	serviceEndpointAfterRoundTrip, projectID, err := expandServiceEndpointDockerRegistry(resourceData)
 
+	require.Equal(t, dockerRegistryTestServiceEndpoint, *serviceEndpointAfterRoundTrip)
+	require.Equal(t, dockerRegistryTestServiceEndpointProjectID, projectID)
 	require.Nil(t, err)
-	require.Equal(t, dhTestServiceEndpoint, *serviceEndpointAfterRoundTrip)
-	require.Equal(t, dhTestServiceEndpointProjectID, projectID)
 }
 
 // verifies that if an error is produced on create, the error is not swallowed
-func TestAzureDevOpsServiceEndpointDockerHub_Create_DoesNotSwallowError(t *testing.T) {
+func TestAzureDevOpsServiceEndpointDockerRegistry_Create_DoesNotSwallowError(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	r := resourceServiceEndpointDockerHub()
+	r := resourceServiceEndpointDockerRegistry()
 	resourceData := schema.TestResourceDataRaw(t, r.Schema, nil)
-	flattenServiceEndpointDockerHub(resourceData, &dhTestServiceEndpoint, dhTestServiceEndpointProjectID)
+	flattenServiceEndpointDockerRegistry(resourceData, &dockerRegistryTestServiceEndpoint, dockerRegistryTestServiceEndpointProjectID)
 
 	buildClient := azdosdkmocks.NewMockServiceendpointClient(ctrl)
 	clients := &config.AggregatedClient{ServiceEndpointClient: buildClient, Ctx: context.Background()}
 
-	expectedArgs := serviceendpoint.CreateServiceEndpointArgs{Endpoint: &dhTestServiceEndpoint, Project: dhTestServiceEndpointProjectID}
+	expectedArgs := serviceendpoint.CreateServiceEndpointArgs{Endpoint: &dockerRegistryTestServiceEndpoint, Project: dockerRegistryTestServiceEndpointProjectID}
 	buildClient.
 		EXPECT().
 		CreateServiceEndpoint(clients.Ctx, expectedArgs).
@@ -87,18 +87,18 @@ func TestAzureDevOpsServiceEndpointDockerHub_Create_DoesNotSwallowError(t *testi
 }
 
 // verifies that if an error is produced on a read, it is not swallowed
-func TestAzureDevOpsServiceEndpointDockerHub_Read_DoesNotSwallowError(t *testing.T) {
+func TestAzureDevOpsServiceEndpointDockerRegistry_Read_DoesNotSwallowError(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	r := resourceServiceEndpointDockerHub()
+	r := resourceServiceEndpointDockerRegistry()
 	resourceData := schema.TestResourceDataRaw(t, r.Schema, nil)
-	flattenServiceEndpointDockerHub(resourceData, &dhTestServiceEndpoint, dhTestServiceEndpointProjectID)
+	flattenServiceEndpointDockerRegistry(resourceData, &dockerRegistryTestServiceEndpoint, dockerRegistryTestServiceEndpointProjectID)
 
 	buildClient := azdosdkmocks.NewMockServiceendpointClient(ctrl)
 	clients := &config.AggregatedClient{ServiceEndpointClient: buildClient, Ctx: context.Background()}
 
-	expectedArgs := serviceendpoint.GetServiceEndpointDetailsArgs{EndpointId: dhTestServiceEndpoint.Id, Project: dhTestServiceEndpointProjectID}
+	expectedArgs := serviceendpoint.GetServiceEndpointDetailsArgs{EndpointId: dockerRegistryTestServiceEndpoint.Id, Project: dockerRegistryTestServiceEndpointProjectID}
 	buildClient.
 		EXPECT().
 		GetServiceEndpointDetails(clients.Ctx, expectedArgs).
@@ -110,18 +110,18 @@ func TestAzureDevOpsServiceEndpointDockerHub_Read_DoesNotSwallowError(t *testing
 }
 
 // verifies that if an error is produced on a delete, it is not swallowed
-func TestAzureDevOpsServiceEndpointDockerHub_Delete_DoesNotSwallowError(t *testing.T) {
+func TestAzureDevOpsServiceEndpointDockerRegistry_Delete_DoesNotSwallowError(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	r := resourceServiceEndpointDockerHub()
+	r := resourceServiceEndpointDockerRegistry()
 	resourceData := schema.TestResourceDataRaw(t, r.Schema, nil)
-	flattenServiceEndpointDockerHub(resourceData, &dhTestServiceEndpoint, dhTestServiceEndpointProjectID)
+	flattenServiceEndpointDockerRegistry(resourceData, &dockerRegistryTestServiceEndpoint, dockerRegistryTestServiceEndpointProjectID)
 
 	buildClient := azdosdkmocks.NewMockServiceendpointClient(ctrl)
 	clients := &config.AggregatedClient{ServiceEndpointClient: buildClient, Ctx: context.Background()}
 
-	expectedArgs := serviceendpoint.DeleteServiceEndpointArgs{EndpointId: dhTestServiceEndpoint.Id, Project: dhTestServiceEndpointProjectID}
+	expectedArgs := serviceendpoint.DeleteServiceEndpointArgs{EndpointId: dockerRegistryTestServiceEndpoint.Id, Project: dockerRegistryTestServiceEndpointProjectID}
 	buildClient.
 		EXPECT().
 		DeleteServiceEndpoint(clients.Ctx, expectedArgs).
@@ -133,21 +133,21 @@ func TestAzureDevOpsServiceEndpointDockerHub_Delete_DoesNotSwallowError(t *testi
 }
 
 // verifies that if an error is produced on an update, it is not swallowed
-func TestAzureDevOpsServiceEndpointDockerHub_Update_DoesNotSwallowError(t *testing.T) {
+func TestAzureDevOpsServiceEndpointDockerRegistry_Update_DoesNotSwallowError(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	r := resourceServiceEndpointDockerHub()
+	r := resourceServiceEndpointDockerRegistry()
 	resourceData := schema.TestResourceDataRaw(t, r.Schema, nil)
-	flattenServiceEndpointDockerHub(resourceData, &dhTestServiceEndpoint, dhTestServiceEndpointProjectID)
+	flattenServiceEndpointDockerRegistry(resourceData, &dockerRegistryTestServiceEndpoint, dockerRegistryTestServiceEndpointProjectID)
 
 	buildClient := azdosdkmocks.NewMockServiceendpointClient(ctrl)
 	clients := &config.AggregatedClient{ServiceEndpointClient: buildClient, Ctx: context.Background()}
 
 	expectedArgs := serviceendpoint.UpdateServiceEndpointArgs{
-		Endpoint:   &dhTestServiceEndpoint,
-		EndpointId: dhTestServiceEndpoint.Id,
-		Project:    dhTestServiceEndpointProjectID,
+		Endpoint:   &dockerRegistryTestServiceEndpoint,
+		EndpointId: dockerRegistryTestServiceEndpoint.Id,
+		Project:    dockerRegistryTestServiceEndpointProjectID,
 	}
 
 	buildClient.
@@ -166,25 +166,27 @@ func TestAzureDevOpsServiceEndpointDockerHub_Update_DoesNotSwallowError(t *testi
 
 // validates that an apply followed by another apply (i.e., resource update) will be reflected in AzDO and the
 // underlying terraform state.
-func TestAccAzureDevOpsServiceEndpointDockerHub_CreateAndUpdate(t *testing.T) {
+func TestAccAzureDevOpsServiceEndpointDockerRegistry_CreateAndUpdate(t *testing.T) {
 	projectName := testhelper.TestAccResourcePrefix + acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	serviceEndpointNameFirst := testhelper.TestAccResourcePrefix + acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	serviceEndpointNameSecond := testhelper.TestAccResourcePrefix + acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
+	// username := "u" + acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
+	// password := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 
-	tfSvcEpNode := "azuredevops_serviceendpoint_dockerhub.serviceendpoint"
+	tfSvcEpNode := "azuredevops_serviceendpoint_dockerregistry.serviceendpoint"
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testhelper.TestAccPreCheck(t, &[]string{
-				"AZDO_DOCKERHUB_SERVICE_CONNECTION_USERNAME",
-				"AZDO_DOCKERHUB_SERVICE_CONNECTION_EMAIL",
-				"AZDO_DOCKERHUB_SERVICE_CONNECTION_PASSWORD",
+				"AZDO_DOCKERREGISTRY_SERVICE_CONNECTION_USERNAME",
+				"AZDO_DOCKERREGISTRY_SERVICE_CONNECTION_EMAIL",
+				"AZDO_DOCKERREGISTRY_SERVICE_CONNECTION_PASSWORD",
 			})
 		},
 		Providers:    testAccProviders,
-		CheckDestroy: testAccServiceEndpointDockerHubCheckDestroy,
+		CheckDestroy: testAccServiceEndpointDockerRegistryCheckDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testhelper.TestAccServiceEndpointDockerHubResource(projectName, serviceEndpointNameFirst),
+				Config: testhelper.TestAccServiceEndpointDockerRegistryResource(projectName, serviceEndpointNameFirst),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttrSet(tfSvcEpNode, "project_id"),
 					resource.TestCheckResourceAttrSet(tfSvcEpNode, "docker_username"),
@@ -192,10 +194,10 @@ func TestAccAzureDevOpsServiceEndpointDockerHub_CreateAndUpdate(t *testing.T) {
 					resource.TestCheckResourceAttr(tfSvcEpNode, "docker_password", ""),
 					resource.TestCheckResourceAttrSet(tfSvcEpNode, "docker_password_hash"),
 					resource.TestCheckResourceAttr(tfSvcEpNode, "service_endpoint_name", serviceEndpointNameFirst),
-					testAccCheckServiceEndpointDockerHubResourceExists(serviceEndpointNameFirst),
+					testAccCheckServiceEndpointDockerRegistryResourceExists(serviceEndpointNameFirst),
 				),
 			}, {
-				Config: testhelper.TestAccServiceEndpointDockerHubResource(projectName, serviceEndpointNameSecond),
+				Config: testhelper.TestAccServiceEndpointDockerRegistryResource(projectName, serviceEndpointNameSecond),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttrSet(tfSvcEpNode, "project_id"),
 					resource.TestCheckResourceAttrSet(tfSvcEpNode, "docker_username"),
@@ -203,7 +205,7 @@ func TestAccAzureDevOpsServiceEndpointDockerHub_CreateAndUpdate(t *testing.T) {
 					resource.TestCheckResourceAttr(tfSvcEpNode, "docker_password", ""),
 					resource.TestCheckResourceAttrSet(tfSvcEpNode, "docker_password_hash"),
 					resource.TestCheckResourceAttr(tfSvcEpNode, "service_endpoint_name", serviceEndpointNameSecond),
-					testAccCheckServiceEndpointDockerHubResourceExists(serviceEndpointNameSecond),
+					testAccCheckServiceEndpointDockerRegistryResourceExists(serviceEndpointNameSecond),
 				),
 			},
 		},
@@ -212,14 +214,14 @@ func TestAccAzureDevOpsServiceEndpointDockerHub_CreateAndUpdate(t *testing.T) {
 
 // Given the name of an AzDO service endpoint, this will return a function that will check whether
 // or not the resource (1) exists in the state and (2) exist in AzDO and (3) has the correct name
-func testAccCheckServiceEndpointDockerHubResourceExists(expectedName string) resource.TestCheckFunc {
+func testAccCheckServiceEndpointDockerRegistryResourceExists(expectedName string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		serviceEndpointDef, ok := s.RootModule().Resources["azuredevops_serviceendpoint_dockerhub.serviceendpoint"]
+		serviceEndpointDef, ok := s.RootModule().Resources["azuredevops_serviceendpoint_dockerregistry.serviceendpoint"]
 		if !ok {
 			return fmt.Errorf("Did not find a service endpoint in the TF state")
 		}
 
-		serviceEndpoint, err := getServiceEndpointDockerHubFromResource(serviceEndpointDef)
+		serviceEndpoint, err := getServiceEndpointDockerRegistryFromResource(serviceEndpointDef)
 		if err != nil {
 			return err
 		}
@@ -234,14 +236,14 @@ func testAccCheckServiceEndpointDockerHubResourceExists(expectedName string) res
 
 // verifies that all service endpoints referenced in the state are destroyed. This will be invoked
 // *after* terrafform destroys the resource but *before* the state is wiped clean.
-func testAccServiceEndpointDockerHubCheckDestroy(s *terraform.State) error {
+func testAccServiceEndpointDockerRegistryCheckDestroy(s *terraform.State) error {
 	for _, resource := range s.RootModule().Resources {
-		if resource.Type != "azuredevops_serviceendpoint_dockerhub" {
+		if resource.Type != "azuredevops_serviceendpoint_dockerregistry" {
 			continue
 		}
 
 		// indicates the service endpoint still exists - this should fail the test
-		if _, err := getServiceEndpointDockerHubFromResource(resource); err == nil {
+		if _, err := getServiceEndpointDockerRegistryFromResource(resource); err == nil {
 			return fmt.Errorf("Unexpectedly found a service endpoint that should be deleted")
 		}
 	}
@@ -250,7 +252,7 @@ func testAccServiceEndpointDockerHubCheckDestroy(s *terraform.State) error {
 }
 
 // given a resource from the state, return a service endpoint (and error)
-func getServiceEndpointDockerHubFromResource(resource *terraform.ResourceState) (*serviceendpoint.ServiceEndpoint, error) {
+func getServiceEndpointDockerRegistryFromResource(resource *terraform.ResourceState) (*serviceendpoint.ServiceEndpoint, error) {
 	serviceEndpointDefID, err := uuid.Parse(resource.Primary.ID)
 	if err != nil {
 		return nil, err